atomify log
